third_party/blink/web_tests/platform/generic/fast/events/focus-event-source-device-from-mouse-expected.txt

Test that focus events from a mouse device should have a sourceCapabilities and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

tests on div
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on input
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on button
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on textarea
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on select
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on a
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


PASS successfullyParsed is true

TEST COMPLETE
